+++ lib/models/2.2.0/asyncapi.js
@@ -5,9 +5,9 @@ const Info = require('./info');
 const Server = require('./server');
 const Channel = require('./channel');
 const Components = require('./components');
-const MixinExternalDocs = require('../mixins/external-docs');
-const MixinTags = require('../mixins/tags');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinExternalDocs = require('./mixins/external-docs');
+const MixinTags = require('./mixins/tags');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 const {xParserCircle, xParserCircleProps} = require('../constants');
 const {assignNameToAnonymousMessages, assignNameToComponentMessages, assignUidToComponentSchemas, assignUidToParameterSchemas, assignIdToAnonymousSchemas} = require('../anonymousNaming');
 const {traverseAsyncApiDocument, SchemaIteratorCallbackType} = require('../iterators');
