+++ lib/models/2.2.0/channel-parameter.js
@@ -3,8 +3,8 @@ const { mix } = require('./utils');
 const Base = require('./base');
 const Schema = require('./schema');
 
-const MixinDescription = require('../mixins/description');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinDescription = require('./mixins/description');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with a ChannelParameter object.
