+++ lib/models/2.2.0/channel.js
@@ -5,9 +5,9 @@ const ChannelParameter = require('./channel-parameter');
 const PublishOperation = require('./publish-operation');
 const SubscribeOperation = require('./subscribe-operation');
 
-const MixinDescription = require('../mixins/description');
-const MixinBindings = require('../mixins/bindings');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinDescription = require('./mixins/description');
+const MixinBindings = require('./mixins/bindings');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with a Channel object.
