+++ lib/models/2.2.0/components.js
@@ -9,7 +9,7 @@ const CorrelationId = require('./correlation-id');
 const OperationTrait = require('./operation-trait');
 const MessageTrait = require('./message-trait');
 
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with a Components object.
