+++ lib/models/2.2.0/contact.js
@@ -3,7 +3,7 @@ const { mix } = require('./utils');
 
 const Base = require('./base');
 
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with the Contact object.
