+++ lib/models/2.2.0/info.js
@@ -4,8 +4,8 @@ const Base = require('./base');
 const License = require('./license');
 const Contact = require('./contact');
 
-const MixinDescription = require('../mixins/description');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinDescription = require('./mixins/description');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with the Info object.
