+++ lib/models/2.2.0/message-trait.js
