+++ lib/models/2.2.0/message-traitable.js
@@ -4,11 +4,11 @@ const Base = require('./base');
 const Schema = require('./schema');
 const CorrelationId = require('./correlation-id');
 
-const MixinDescription = require('../mixins/description');
-const MixinExternalDocs = require('../mixins/external-docs');
-const MixinTags = require('../mixins/tags');
-const MixinBindings = require('../mixins/bindings');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinDescription = require('./mixins/description');
+const MixinExternalDocs = require('./mixins/external-docs');
+const MixinTags = require('./mixins/tags');
+const MixinBindings = require('./mixins/bindings');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with a the common properties that Message and MessageTrait objects have.
