+++ lib/models/2.2.0/mixins/description.js
