+++ lib/models/2.2.0/oauth-flow.js
@@ -2,7 +2,7 @@ const { mix } = require('./utils');
 
 const Base = require('./base');
 
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with a OAuthFlow object.
