+++ lib/models/2.2.0/operation-trait.js
