+++ lib/models/2.2.0/operation-traitable.js
@@ -2,11 +2,11 @@ const { mix } = require('./utils');
 
 const Base = require('./base');
 
-const MixinDescription = require('../mixins/description');
-const MixinTags = require('../mixins/tags');
-const MixinExternalDocs = require('../mixins/external-docs');
-const MixinBindings = require('../mixins/bindings');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinDescription = require('.././mixins/description');
+const MixinTags = require('.././mixins/tags');
+const MixinExternalDocs = require('.././mixins/external-docs');
+const MixinBindings = require('.././mixins/bindings');
+const MixinSpecificationExtensions = require('.././mixins/specification-extensions');
 
 /**
  * Implements functions to deal with the common properties Operation and OperationTrait object have.
