+++ lib/models/2.2.0/publish-operation.js
