+++ lib/models/2.2.0/schema.js
@@ -2,9 +2,9 @@ const { createMapOfType, getMapValueOfType, mix } = require('./utils');
 
 const Base = require('./base');
 
-const MixinDescription = require('../mixins/description');
-const MixinExternalDocs = require('../mixins/external-docs');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinDescription = require('./mixins/description');
+const MixinExternalDocs = require('./mixins/external-docs');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with a Schema object.
