+++ lib/models/2.2.0/security-scheme.js
@@ -3,8 +3,8 @@ const { createMapOfType, mix } = require('./utils');
 const Base = require('./base');
 const OAuthFlow = require('./oauth-flow');
 
-const MixinDescription = require('../mixins/description');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinDescription = require('./mixins/description');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with a SecurityScheme object.
