+++ lib/models/2.2.0/server-variable.js
@@ -2,8 +2,8 @@ const { mix } = require('./utils');
 
 const Base = require('./base');
 
-const MixinDescription = require('../mixins/description');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinDescription = require('./mixins/description');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with a ServerVariable object.
