+++ lib/models/2.2.0/server.js
@@ -4,9 +4,9 @@ const Base = require('./base');
 const ServerVariable = require('./server-variable');
 const ServerSecurityRequirement = require('./server-security-requirement');
 
-const MixinDescription = require('../mixins/description');
-const MixinBindings = require('../mixins/bindings');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinDescription = require('./mixins/description');
+const MixinBindings = require('./mixins/bindings');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with a Server object.
