+++ lib/models/2.2.0/tag.js
@@ -2,9 +2,9 @@ const { mix } = require('./utils');
 
 const Base = require('./base');
 
-const MixinDescription = require('../mixins/description');
-const MixinExternalDocs = require('../mixins/external-docs');
-const MixinSpecificationExtensions = require('../mixins/specification-extensions');
+const MixinDescription = require('./mixins/description');
+const MixinExternalDocs = require('./mixins/external-docs');
+const MixinSpecificationExtensions = require('./mixins/specification-extensions');
 
 /**
  * Implements functions to deal with a Tag object.
