+++ lib/models/intents/base.js
