+++ lib/models/intents/channel-parameter.js
