+++ lib/models/intents/channel.js
