+++ lib/models/intents/external-document.js
