+++ lib/models/intents/info.js
