+++ lib/models/intents/license.js
