+++ lib/models/intents/operation.js
